+++ crawler/src/main/scala-2.11/crawler_run.scala
@@ -2,5 +2,5 @@ import wikicrawler.WikiCrawler
 
 object crawler_run extends App {
   val entry_point = "https://en.wikipedia.org/wiki/Information_retrieval"
-  WikiCrawler.start("https://en.wikipedia.org/wiki/463")
+  new WikiCrawler().start(entry_point)
 }
